// packages/js-commons/src/enums.ts
export enum AppCategory {
  Demo = "Demo",
  Template = "Template",
}

export enum AccessType {
  Read = "read",
  Write = "write",
  Owner = "owner",
}

export enum EnsembleDocumentType {
  Font = "font",
  I18n = "i18n",
  Theme = "theme",
  Asset = "asset",
  Label = "label",
  Screen = "screen",
  Secrets = "secrets",
  Environment = "config",
  Widget = "internal_widget",
  Script = "internal_script",
}

export enum CollectionsName {
  Apps = "apps",
  Users = "users",
  Labels = "labels",
  History = "history",
  Artifacts = "artifacts",
  InternalArtifacts = "internal_artifacts",
}